backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.9.2
attrs==25.3.0
beautifulsoup4==4.12.3
billiard==4.2.4
cachetools==5.5.2
celery==5.6.2
certifi==2026.2.25
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
distro==1.9.0
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.3.0
django-filter==25.1
django-timezone-field==7.2.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
google-ai-generativelanguage==0.6.5
google-api-core==2.30.0
google-api-python-client==2.118.0
google-auth==2.36.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.13.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.2
lxml==5.1.0
Markdown==3.9
oauthlib==3.3.1
openai==2.24.0
packaging==26.0
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==4.25.8
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.8.0
pyparsing==3.3.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
sentry-sdk==2.53.0
setuptools==82.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sqlparse==0.5.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
